+++ embassy-lora/src/lib.rs
@@ -10,10 +10,13 @@ pub(crate) mod fmt;
 pub mod iv;
 
 #[cfg(feature = "stm32wl")]
+#[deprecated(note = "use the external LoRa physical layer crate - https://crates.io/crates/lora-phy")]
 pub mod stm32wl;
 #[cfg(feature = "sx126x")]
+#[deprecated(note = "use the external LoRa physical layer crate - https://crates.io/crates/lora-phy")]
 pub mod sx126x;
 #[cfg(feature = "sx127x")]
+#[deprecated(note = "use the external LoRa physical layer crate - https://crates.io/crates/lora-phy")]
 pub mod sx127x;
 
 #[cfg(feature = "time")]
