+++ examples/rp/src/bin/lora_p2p_receive.rs
@@ -1,4 +1,4 @@
-//! This example runs on the STM32 LoRa Discovery board, which has a builtin Semtech Sx1276 radio.
+//! This example runs on the Raspberry Pi Pico with a Waveshare board containing a Semtech Sx1262 radio.
 //! It demonstrates LORA P2P receive functionality in conjunction with the lora_p2p_send example.
 #![no_std]
 #![no_main]
@@ -7,14 +7,12 @@
 
 use defmt::*;
 use embassy_executor::Spawner;
-use embassy_lora::iv::Stm32l0InterfaceVariant;
-use embassy_stm32::exti::{Channel, ExtiInput};
-use embassy_stm32::gpio::{Input, Level, Output, Pin, Pull, Speed};
-use embassy_stm32::spi;
-use embassy_stm32::time::khz;
+use embassy_lora::iv::GenericSx126xInterfaceVariant;
+use embassy_rp::gpio::{Input, Level, Output, Pin, Pull};
+use embassy_rp::spi::{Config, Spi};
 use embassy_time::{Delay, Duration, Timer};
 use lora_phy::mod_params::*;
-use lora_phy::sx1276_7_8_9::SX1276_7_8_9;
+use lora_phy::sx1261_2::SX1261_2;
 use lora_phy::LoRa;
 use {defmt_rtt as _, panic_probe as _};
 
@@ -22,35 +20,30 @@ const LORA_FREQUENCY_IN_HZ: u32 = 903_900_000; // warning: set this appropriatel
 
 #[embassy_executor::main]
 async fn main(_spawner: Spawner) {
-    let mut config = embassy_stm32::Config::default();
-    config.rcc.mux = embassy_stm32::rcc::ClockSrc::HSI16;
-    config.rcc.enable_hsi48 = true;
-    let p = embassy_stm32::init(config);
+    let p = embassy_rp::init(Default::default());
 
-    // SPI for sx1276
-    let spi = spi::Spi::new(
-        p.SPI1,
-        p.PB3,
-        p.PA7,
-        p.PA6,
-        p.DMA1_CH3,
-        p.DMA1_CH2,
-        khz(200),
-        spi::Config::default(),
-    );
+    let miso = p.PIN_12;
+    let mosi = p.PIN_11;
+    let clk = p.PIN_10;
+    let spi = Spi::new(p.SPI1, clk, mosi, miso, p.DMA_CH0, p.DMA_CH1, Config::default());
 
-    let nss = Output::new(p.PA15.degrade(), Level::High, Speed::Low);
-    let reset = Output::new(p.PC0.degrade(), Level::High, Speed::Low);
+    let nss = Output::new(p.PIN_3.degrade(), Level::High);
+    let reset = Output::new(p.PIN_15.degrade(), Level::High);
+    let dio1 = Input::new(p.PIN_20.degrade(), Pull::None);
+    let busy = Input::new(p.PIN_2.degrade(), Pull::None);
 
-    let irq_pin = Input::new(p.PB4.degrade(), Pull::Up);
-    let irq = ExtiInput::new(irq_pin, p.EXTI4.degrade());
-
-    let iv = Stm32l0InterfaceVariant::new(nss, reset, irq, None, None).unwrap();
+    let iv = GenericSx126xInterfaceVariant::new(nss, reset, dio1, busy, None, None).unwrap();
 
     let mut delay = Delay;
 
     let mut lora = {
-        match LoRa::new(SX1276_7_8_9::new(BoardType::Stm32l0Sx1276, spi, iv), false, &mut delay).await {
+        match LoRa::new(
+            SX1261_2::new(BoardType::RpPicoWaveshareSx1262, spi, iv),
+            false,
+            &mut delay,
+        )
+        .await
+        {
             Ok(l) => l,
             Err(err) => {
                 info!("Radio error = {}", err);
@@ -59,12 +52,7 @@ async fn main(_spawner: Spawner) {
         }
     };
 
-    let mut debug_indicator = Output::new(p.PB5, Level::Low, Speed::Low);
-    let mut start_indicator = Output::new(p.PB6, Level::Low, Speed::Low);
-
-    start_indicator.set_high();
-    Timer::after(Duration::from_secs(5)).await;
-    start_indicator.set_low();
+    let mut debug_indicator = Output::new(p.PIN_25, Level::Low);
 
     let mut receiving_buffer = [00u8; 100];
