+++ examples/stm32wl/src/bin/lora_p2p_receive.rs
@@ -1,5 +1,5 @@
-//! This example runs on a STM32WL board, which has a builtin Semtech Sx1262 radio.
-//! It demonstrates LORA P2P send functionality.
+//! This example runs on the STM32WL board, which has a builtin Semtech Sx1262 radio.
+//! It demonstrates LORA P2P receive functionality in conjunction with the lora_p2p_send example.
 #![no_std]
 #![no_main]
 #![macro_use]
@@ -17,7 +17,7 @@ use embassy_stm32::rcc::low_level::RccPeripheral;
 use embassy_stm32::spi::{BitOrder, Config as SpiConfig, Spi, MODE_0};
 use embassy_stm32::time::Hertz;
 use embassy_stm32::{interrupt, into_ref, Peripheral};
-use embassy_time::Delay;
+use embassy_time::{Delay, Duration, Timer};
 use lora_phy::mod_params::*;
 use lora_phy::sx1261_2::SX1261_2;
 use lora_phy::LoRa;
@@ -59,6 +59,15 @@ async fn main(_spawner: Spawner) {
         }
     };
 
+    let mut debug_indicator = Output::new(p.PB9, Level::Low, Speed::Low);
+    let mut start_indicator = Output::new(p.PB15, Level::Low, Speed::Low);
+
+    start_indicator.set_high();
+    Timer::after(Duration::from_secs(5)).await;
+    start_indicator.set_low();
+
+    let mut receiving_buffer = [00u8; 100];
+
     let mdltn_params = {
         match lora.create_modulation_params(
             SpreadingFactor::_10,
@@ -74,8 +83,8 @@ async fn main(_spawner: Spawner) {
         }
     };
 
-    let mut tx_pkt_params = {
-        match lora.create_tx_packet_params(4, false, true, false, &mdltn_params) {
+    let rx_pkt_params = {
+        match lora.create_rx_packet_params(4, false, receiving_buffer.len() as u8, true, false, &mdltn_params) {
             Ok(pp) => pp,
             Err(err) => {
                 info!("Radio error = {}", err);
@@ -84,7 +93,10 @@ async fn main(_spawner: Spawner) {
         }
     };
 
-    match lora.prepare_for_tx(&mdltn_params, 20, false).await {
+    match lora
+        .prepare_for_rx(&mdltn_params, &rx_pkt_params, None, true, false, 0, 0x00ffffffu32)
+        .await
+    {
         Ok(()) => {}
         Err(err) => {
             info!("Radio error = {}", err);
@@ -92,19 +104,24 @@ async fn main(_spawner: Spawner) {
         }
     };
 
-    let buffer = [0x01u8, 0x02u8, 0x03u8];
-    match lora.tx(&mdltn_params, &mut tx_pkt_params, &buffer, 0xffffff).await {
-        Ok(()) => {
-            info!("TX DONE");
+    loop {
+        receiving_buffer = [00u8; 100];
+        match lora.rx(&rx_pkt_params, &mut receiving_buffer).await {
+            Ok((received_len, _rx_pkt_status)) => {
+                if (received_len == 3)
+                    && (receiving_buffer[0] == 0x01u8)
+                    && (receiving_buffer[1] == 0x02u8)
+                    && (receiving_buffer[2] == 0x03u8)
+                {
+                    info!("rx successful");
+                    debug_indicator.set_high();
+                    Timer::after(Duration::from_secs(5)).await;
+                    debug_indicator.set_low();
+                } else {
+                    info!("rx unknown packet");
+                }
+            }
+            Err(err) => info!("rx unsuccessful = {}", err),
         }
-        Err(err) => {
-            info!("Radio error = {}", err);
-            return;
-        }
-    };
-
-    match lora.sleep(&mut delay).await {
-        Ok(()) => info!("Sleep successful"),
-        Err(err) => info!("Sleep unsuccessful = {}", err),
     }
 }
